Differentiate between trees and forest.
In the previous version of the code, a node would point to other nodes on the
memo. That meant that if we populated a class in the memo, the nodes pointing to
any of that class's nodes would not see the new alternative plans.

To address that, we now have a tree data type. Following the common terminology
in the literature, we call that type an 'expr'. It isn't an expression,
really. It is a forest of expressions that all share a same root operator on the
memo.  In other words, The 'left' and 'right' pointers out of an 'expr' point to
a memo class.

We still kept the 'node' type around, which is our tree. In a number of
occasions we want to insert or extract a tree from the memo, so that type is
still very useful. Note also that we want to test and apply transformations at
the tree level, not at the forest one. So, again, trees are very much not going
anywhere.

If the terminology 'node' => tree and 'expr' => forest sounds confusing, it is
because it is. We did not want to change the code too much so that people
following opttoy could see better the steps taken here. But we'll soon need to
do something about it.