Prepare the code for a larger number of transformations.
The search strategy now does not call transformations directly. Instead, it asks
the root node of an expression which transformations it is compatible with --
and try to apply those.

We went out of the way here to keep a familiar code structure. But we clearly
went beyond the limits of good software engineering... Luckily, this code is
simply a discussion generator.