Establish a way to test if a node belongs to a Memo class.
We're currently equating a node ID with its Memo class. This creates problems if
we try to enumerate the joins of four or more tables. The problem is that the
join associativity transformation creates an intermediate join that gets
misplaced on the Memo.

The solution here is to separate ID from the node's Memo Class, which we call
"equivalent class". We're not using this throughout but only on the critical
region of the code, where we decide where to place a new node on the Memo.

The code deserve to be cleaned up on several counts but I thought it's easier to
start a cleanup if the code is correct.