Adjust logical equivalency for expressions
Previously, logical equivalency for expressions was defined as
equivalent relational properties. But this is too high of a bar to reach
as we need to ensure that the relational properties encode all of the
differences between expressions. Now we invert the default and assume
two expressions are not equivalent without specific handling. In
particular, this change only allows logical equivalency between
expressions when adding transformed expressions back into the memo and
between inner-join expressions (see `memoExpr.groupFingerprint()`).