Add functions to filter a histogram
Add functions to filter a histogram given a predicate such as
x < 4, in which the column is compared to a constant with a
scalar operator.  The new functions filter the histogram for
the operators ltOp, leOp, gtOp and geOp. The result is a histogram
which essentially has a subset of the buckets in the original
histogram. There is some complexity if the constant in the predicate
falls in the middle of a bucket, because then the bucket
must be split into two buckets.