Every scan generates a new set of output cols
Previously, the output cols for a scan of a table were the same for
every time the table was referenced in a query. But that causes badness
for pushing down filters in self-joins such as:

  SELECT * FROM a AS l JOIN a AS r ON (l.x = r.y)

For this query, `l.x` is not equivalent to `r.x` and `l.y` is not
equivalent to `r.y`.